Path conditions modeling:     - Constraints classes moved to a constraint package
