+++ client/src/main/java/org/evosuite/symbolic/expr/constraint/ConstraintVisitor.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with EvoSuite. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.evosuite.symbolic.expr;
+package org.evosuite.symbolic.expr.constraint;
 
 public interface ConstraintVisitor<K,V> {
