+++ client/src/main/java/org/evosuite/symbolic/solver/cvc4/NonLinearConstraintVisitor.java
@@ -19,10 +19,10 @@
  */
 package org.evosuite.symbolic.solver.cvc4;
 
-import org.evosuite.symbolic.expr.ConstraintVisitor;
-import org.evosuite.symbolic.expr.IntegerConstraint;
-import org.evosuite.symbolic.expr.RealConstraint;
-import org.evosuite.symbolic.expr.StringConstraint;
+import org.evosuite.symbolic.expr.constraint.ConstraintVisitor;
+import org.evosuite.symbolic.expr.constraint.IntegerConstraint;
+import org.evosuite.symbolic.expr.constraint.RealConstraint;
+import org.evosuite.symbolic.expr.constraint.StringConstraint;
 
 final class NonLinearConstraintVisitor implements ConstraintVisitor<Boolean, Void> {
