+++ client/src/main/java/org/evosuite/symbolic/vm/HeapVM.java
@@ -25,7 +25,7 @@ import java.lang.reflect.Array;
 import java.lang.reflect.Field;
 
 import org.evosuite.symbolic.expr.Expression;
-import org.evosuite.symbolic.expr.IntegerConstraint;
+import org.evosuite.symbolic.expr.constraint.IntegerConstraint;
 import org.evosuite.symbolic.expr.bv.IntegerConstant;
 import org.evosuite.symbolic.expr.bv.IntegerValue;
 import org.evosuite.symbolic.expr.fp.RealValue;
