+++ client/src/main/java/org/evosuite/symbolic/vm/JumpVM.java
@@ -24,7 +24,7 @@ import java.util.Vector;
 import org.evosuite.symbolic.expr.bv.IntegerConstant;
 import org.evosuite.symbolic.expr.bv.IntegerValue;
 import org.evosuite.symbolic.expr.ref.ReferenceExpression;
-import org.evosuite.symbolic.expr.IntegerConstraint;
+import org.evosuite.symbolic.expr.constraint.IntegerConstraint;
 
 import org.evosuite.dse.AbstractVM;
