+++ client/src/main/java/org/evosuite/symbolic/vm/SymbolicFunction.java
@@ -20,7 +20,7 @@
 package org.evosuite.symbolic.vm;
 
 import org.objectweb.asm.Type;
-import org.evosuite.symbolic.expr.IntegerConstraint;
+import org.evosuite.symbolic.expr.constraint.IntegerConstraint;
 import org.evosuite.symbolic.expr.bv.IntegerValue;
 import org.evosuite.symbolic.expr.fp.RealValue;
 import org.evosuite.symbolic.expr.ref.ReferenceConstant;
