+++ client/src/main/java/org/evosuite/symbolic/vm/string/buffer/StringBuffer_Append.java
@@ -19,7 +19,7 @@
  */
 package org.evosuite.symbolic.vm.string.buffer;
 
-import org.evosuite.symbolic.expr.IntegerConstraint;
+import org.evosuite.symbolic.expr.constraint.IntegerConstraint;
 import org.evosuite.symbolic.expr.Operator;
 import org.evosuite.symbolic.expr.bv.IntegerValue;
 import org.evosuite.symbolic.expr.fp.RealValue;
